migrate ps cmd
